requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
phonenumbers==8.13.36
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
